# contracts/openapi/operations.v1.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "Cutover operations API",
    "version": "1.0.0",
    "description": "Local synthetic operations. Every request checks role and site. Errors use RFC 9457: 400 malformed, 401 unauthenticated, 403 role denied, 404 outside site or absent, 409 conflict, 413 body over 64 KiB, 422 domain validation, 429 rate limit, 503 durable admission unavailable with Retry-After."
  },
  "servers": [
    {
      "url": "http://localhost:8780"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/api/v1/sites/{siteId}/return-receipts": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "maxLength": 40
          }
        }
      ],
      "post": {
        "operationId": "createReturnReceipts",
        "summary": "Durably accept synthetic work; acceptance is not completion",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/receipt-request"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Committed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/accepted"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/problem"
          }
        }
      },
      "get": {
        "operationId": "listReturnReceipts",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Site-scoped current returns projection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/receipt-page"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    },
    "/api/v1/sites/{siteId}/return-receipts/{id}": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "maxLength": 40
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getReturnReceipts",
        "responses": {
          "200": {
            "description": "Site-scoped current returns projection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/receipt-view"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    },
...
    "/api/v1/sites/{siteId}/return-counters": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "maxLength": 40
          }
        }
      ],
      "get": {
        "operationId": "getReturnCounters",
        "responses": {
          "200": {
            "description": "Site-scoped current returns projection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/return-counters"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    },
    "/api/v1/sites/{siteId}/return-tasks": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "maxLength": 40
          }
        }
      ],
      "get": {
        "operationId": "getReturnTasks",
        "responses": {
          "200": {
            "description": "Site-scoped current returns projection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/return-task-list"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    },
    "/api/v1/sites/{siteId}/return-tasks/{id}/recovery": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "recoverReturnTask",
        "description": "Supervisor resumes an exhausted execution transport investigation with expected version and reason. No allocation or physical command identity is replaced.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../schemas/reconciliation-request.v1.json"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Audited status investigation requested"
          },
          "default": {
            "$ref": "#/components/responses/problem"
          }
        },
        "summary": "Resume an exhausted returns transport investigation; physical uncertainty remains under adapter reconciliation"
      }
    },
    "/api/v1/sites/{siteId}/audit/{owner}": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "maxLength": 40
          }
        },
        {
          "name": "owner",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "core",
              "adapter",
              "execution",
              "returns"
            ]
          }
        },
        {
          "name": "limit",
          "in": "query",
          "required": false,
          "schema": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100,
            "default": 25
          }
        },
        {
          "name": "cursor",
          "in": "query",
          "required": false,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "resource",
          "in": "query",
          "required": false,
          "schema": {
            "type": "string",
            "maxLength": 128
          }
        }
      ],
      "get": {
        "summary": "Read newest-first, site-scoped audit records from the selected service database. Internal diagnostic detail is omitted.",
        "responses": {
          "200": {
            "description": "Read newest-first, site-scoped audit records from the selected service database. Internal diagnostic detail is omitted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/audit-page"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/problem"
          },
          "403": {
            "$ref": "#/components/responses/problem"
          },
          "404": {
            "$ref": "#/components/responses/problem"
          },
          "422": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    },
...
    "/api/v1/sites/{siteId}/return-tasks/{id}": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "maxLength": 40
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "Read one task from its owning service, scoped to the authenticated site.",
        "responses": {
          "200": {
            "description": "Current owner task projection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/return-task-view"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "receipt-request": {
        "$ref": "../schemas/receipt-request.v1.json"
      },
...
      "accepted": {
        "type": "object",
        "required": [
          "id",
          "statusUrl"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "statusUrl": {
            "type": "string"
          }
        }
      },
...
      "receipt-view": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "siteId": {
            "type": "string"
          },
          "externalReceiptRef": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "REGISTERED",
              "SORTING",
              "COMPLETED",
              "RECONCILIATION_REQUIRED"
            ]
          },
          "version": {
            "type": "integer",
            "minimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "observedAt": {
            "type": "string",
            "format": "date-time"
          },
          "counts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "classification": {
                  "type": "string",
                  "enum": [
                    "REUSABLE",
                    "NEEDS_CLEANING",
                    "DAMAGED"
                  ]
                },
                "received": {
                  "type": "integer",
                  "minimum": 0
                },
                "sorted": {
                  "type": "integer",
                  "minimum": 0
                }
              },
              "required": [
                "classification",
                "received",
                "sorted"
              ],
              "additionalProperties": true
            },
            "minItems": 3,
            "maxItems": 3
          },
          "movements": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "movementId": {
                  "type": "string",
                  "format": "uuid"
                },
                "classification": {
                  "type": "string",
                  "enum": [
                    "REUSABLE",
                    "NEEDS_CLEANING",
                    "DAMAGED"
                  ]
                },
                "state": {
                  "type": "string"
                },
                "movement": {
                  "type": "object",
                  "properties": {
                    "zoneId": {
                      "const": "returns",
                      "type": "string"
                    },
                    "quantity": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "destination": {
                      "type": "string",
                      "enum": [
                        "reusable",
                        "cleaning",
                        "damaged"
                      ]
                    },
                    "source": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "zoneId",
                    "quantity",
                    "destination",
                    "source"
                  ],
                  "additionalProperties": true
                },
                "taskId": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "uuid"
                },
                "taskState": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "taskVersion": {
                  "type": [
                    "integer",
                    "null"
                  ]
                },
                "transportFailures": {
                  "type": [
                    "integer",
                    "null"
                  ]
                },
                "transportPaused": {
                  "type": [
                    "boolean",
                    "null"
                  ]
                },
                "lastError": {
                  "type": [
                    "string",
                    "null"
                  ]
                }
              },
              "required": [
                "movementId",
                "classification",
                "state",
                "movement",
                "taskId",
                "taskState",
                "taskVersion",
                "transportFailures",
                "transportPaused",
                "lastError"
              ],
              "additionalProperties": true
            },
            "minItems": 1,
            "maxItems": 3
          }
        },
        "required": [
          "id",
          "siteId",
          "externalReceiptRef",
          "state",
          "version",
          "createdAt",
          "completedAt",
          "observedAt",
          "counts",
          "movements"
        ],
        "additionalProperties": true
      },
      "receipt-page": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "siteId": {
                  "type": "string"
                },
                "externalReceiptRef": {
                  "type": "string"
                },
                "state": {
                  "type": "string",
                  "enum": [
                    "REGISTERED",
                    "SORTING",
                    "COMPLETED",
                    "RECONCILIATION_REQUIRED"
                  ]
                },
                "version": {
                  "type": "integer",
                  "minimum": 0
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "completedAt": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                },
                "observedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "counts": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "classification": {
                        "type": "string",
                        "enum": [
                          "REUSABLE",
                          "NEEDS_CLEANING",
                          "DAMAGED"
                        ]
                      },
                      "received": {
                        "type": "integer",
                        "minimum": 0
                      },
                      "sorted": {
                        "type": "integer",
                        "minimum": 0
                      }
                    },
                    "required": [
                      "classification",
                      "received",
                      "sorted"
                    ],
                    "additionalProperties": true
                  },
                  "minItems": 3,
                  "maxItems": 3
                },
                "movements": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "movementId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "classification": {
                        "type": "string",
                        "enum": [
                          "REUSABLE",
                          "NEEDS_CLEANING",
                          "DAMAGED"
                        ]
                      },
                      "state": {
                        "type": "string"
                      },
                      "movement": {
                        "type": "object",
                        "properties": {
                          "zoneId": {
                            "const": "returns",
                            "type": "string"
                          },
                          "quantity": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "destination": {
                            "type": "string",
                            "enum": [
                              "reusable",
                              "cleaning",
                              "damaged"
                            ]
                          },
                          "source": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "zoneId",
                          "quantity",
                          "destination",
                          "source"
                        ],
                        "additionalProperties": true
                      },
                      "taskId": {
                        "type": [
                          "string",
                          "null"
                        ],
                        "format": "uuid"
                      },
                      "taskState": {
                        "type": [
                          "string",
                          "null"
                        ]
                      },
                      "taskVersion": {
                        "type": [
                          "integer",
                          "null"
                        ]
                      },
                      "transportFailures": {
                        "type": [
                          "integer",
                          "null"
                        ]
                      },
                      "transportPaused": {
                        "type": [
                          "boolean",
                          "null"
                        ]
                      },
                      "lastError": {
                        "type": [
                          "string",
                          "null"
                        ]
                      }
                    },
                    "required": [
                      "movementId",
                      "classification",
                      "state",
                      "movement",
                      "taskId",
                      "taskState",
                      "taskVersion",
                      "transportFailures",
                      "transportPaused",
                      "lastError"
                    ],
                    "additionalProperties": true
                  },
                  "minItems": 1,
                  "maxItems": 3
                }
              },
              "required": [
                "id",
                "siteId",
                "externalReceiptRef",
                "state",
                "version",
                "createdAt",
                "completedAt",
                "observedAt",
                "counts",
                "movements"
              ],
              "additionalProperties": true
            },
            "maxItems": 100
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "observedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "items",
          "nextCursor",
          "observedAt"
        ],
        "additionalProperties": true
      },
      "return-counters": {
        "type": "object",
        "properties": {
          "siteId": {
            "type": "string"
          },
          "observedAt": {
            "type": "string",
            "format": "date-time"
          },
          "classifications": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "classification": {
                  "type": "string",
                  "enum": [
                    "REUSABLE",
                    "NEEDS_CLEANING",
                    "DAMAGED"
                  ]
                },
                "received": {
                  "type": "integer",
                  "minimum": 0
                },
                "sorted": {
                  "type": "integer",
                  "minimum": 0
                },
                "outstanding": {
                  "type": "integer",
                  "minimum": 0
                }
              },
              "required": [
                "classification",
                "received",
                "sorted",
                "outstanding"
              ],
              "additionalProperties": true
            },
            "minItems": 3,
            "maxItems": 3
          }
        },
        "required": [
          "siteId",
          "observedAt",
          "classifications"
        ],
        "additionalProperties": true
      },
      "return-task-list": {
        "type": "array",
        "maxItems": 100,
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid"
            },
            "movementId": {
              "type": "string",
              "format": "uuid"
            },
            "allocationId": {
              "type": "string",
              "format": "uuid"
            },
            "receiptId": {
              "type": "string",
              "format": "uuid"
            },
            "classification": {
              "type": "string",
              "enum": [
                "REUSABLE",
                "NEEDS_CLEANING",
                "DAMAGED"
              ]
            },
            "zoneId": {
              "type": "string",
              "const": "returns"
            },
            "state": {
              "type": "string",
              "enum": [
                "READY",
                "BLOCKED",
                "DISPATCH_REQUESTED",
                "IN_PROGRESS",
                "COMPLETED",
                "RECONCILIATION_REQUIRED"
              ]
            },
            "owner": {
              "type": "string",
              "const": "returns-service"
            },
            "epoch": {
              "type": "integer",
              "minimum": 0
            },
            "version": {
              "type": "integer",
              "minimum": 0
            },
            "lastError": {
              "type": [
                "string",
                "null"
              ]
            },
            "transportFailures": {
              "type": "integer",
              "minimum": 0
            },
            "transportPaused": {
              "type": "boolean"
            },
            "dispatchAcceptedAt": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          "required": [
            "id",
            "movementId",
            "allocationId",
            "receiptId",
            "classification",
            "zoneId",
            "state",
            "owner",
            "epoch",
            "version",
            "lastError",
            "transportFailures",
            "transportPaused",
            "dispatchAcceptedAt"
          ],
          "additionalProperties": true
        }
      },
      "audit-page": {
        "type": "object",
        "required": [
          "owner",
          "observedAt",
          "items",
          "nextCursor"
        ],
        "properties": {
          "owner": {
            "type": "string",
            "enum": [
              "core",
              "adapter",
              "execution",
              "returns"
            ]
          },
          "observedAt": {
            "type": "string",
            "format": "date-time"
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "items": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "type": "object",
              "required": [
                "id",
                "actor",
                "action",
                "resourceId",
                "reason",
                "beforeVersion",
                "afterVersion",
                "outcome",
                "occurredAt"
              ],
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "actor": {
                  "type": "string"
                },
                "action": {
                  "type": "string"
                },
                "resourceId": {
                  "type": "string"
                },
                "reason": {
                  "type": "string"
                },
                "outcome": {
                  "type": "string"
                },
                "beforeVersion": {
                  "type": [
                    "integer",
                    "null"
                  ]
                },
                "afterVersion": {
                  "type": [
                    "integer",
                    "null"
                  ]
                },
                "occurredAt": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "additionalProperties": true
            }
          }
        },
        "additionalProperties": true
      },
...
      "return-task-view": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "movementId": {
            "type": "string",
            "format": "uuid"
          },
          "allocationId": {
            "type": "string",
            "format": "uuid"
          },
          "receiptId": {
            "type": "string",
            "format": "uuid"
          },
          "classification": {
            "type": "string",
            "enum": [
              "REUSABLE",
              "NEEDS_CLEANING",
              "DAMAGED"
            ]
          },
          "zoneId": {
            "type": "string",
            "const": "returns"
          },
          "state": {
            "type": "string",
            "enum": [
              "READY",
              "BLOCKED",
              "DISPATCH_REQUESTED",
              "IN_PROGRESS",
              "COMPLETED",
              "RECONCILIATION_REQUIRED"
            ]
          },
          "owner": {
            "type": "string",
            "const": "returns-service"
          },
          "epoch": {
            "type": "integer",
            "minimum": 0
          },
          "version": {
            "type": "integer",
            "minimum": 0
          },
          "lastError": {
            "type": [
              "string",
              "null"
            ]
          },
          "transportFailures": {
            "type": "integer",
            "minimum": 0
          },
          "transportPaused": {
            "type": "boolean"
          },
          "dispatchAcceptedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "movementId",
          "allocationId",
          "receiptId",
          "classification",
          "zoneId",
          "state",
          "owner",
          "epoch",
          "version",
          "lastError",
          "transportFailures",
          "transportPaused",
          "dispatchAcceptedAt"
        ],
        "additionalProperties": true
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "problem": {
        "description": "Bounded problem response without internal exceptions",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/problem"
            }
          }
        }
      }
    }
  }
}